asdf
asdf

+++ src/com/core/struct32/type32/Elf32_Half.java
@@ -7,11 +7,11 @@ import com.core.util.ConvertUtil;
 
 public class Elf32_Half extends Elf32_StandardType{
 
-	public Elf32_Half(byte[] value,Integer offset){
+	public Elf32_Half(int[] value,Integer offset){
 		super(2,value,offset);
 	}
 	
-	public Elf32_Half(byte[] value,Elf32_Off offset){
+	public Elf32_Half(int[] value,Elf32_Off offset){
 		super(2,value,ConvertUtil.byteArrayToInt(offset.getValue(),MachineCharacter.LITTLE_ENDIAN));
 	}
