+++ src/com/core/struct32/type32/Elf32_Ident.java
@@ -7,11 +7,11 @@ import com.core.util.ConvertUtil;
 
 public class Elf32_Ident extends Elf32_StandardType{
 	
-	public Elf32_Ident(byte[] value,Integer offset){
+	public Elf32_Ident(int[] value,Integer offset){
 		super(16,value,offset);
 	}
 	
-	public Elf32_Ident(byte[] value,Elf32_Off offset){
+	public Elf32_Ident(int[] value,Elf32_Off offset){
 		super(16,value,ConvertUtil.byteArrayToInt(offset.getValue(),MachineCharacter.LITTLE_ENDIAN));
 	}
