+++ src/com/core/struct32/type32/Elf32_Sword.java
@@ -7,11 +7,11 @@ import com.core.util.ConvertUtil;
 
 public class Elf32_Sword extends Elf32_StandardType{
 	
-	public Elf32_Sword(byte[] value,Integer offset){
+	public Elf32_Sword(int[] value,Integer offset){
 		super(4,value,offset);
 	}
 	
-	public Elf32_Sword(byte[] value,Elf32_Off offset){
+	public Elf32_Sword(int[] value,Elf32_Off offset){
 		super(4,value,ConvertUtil.byteArrayToInt(offset.getValue(),MachineCharacter.LITTLE_ENDIAN));
 	}
