+++ src/com/core/struct32/type32/base/Elf32_Type.java
@@ -3,11 +3,11 @@ package com.core.struct32.type32.base;
 public abstract class Elf32_Type {
 
 	private int size;
-	private byte value[];
+	private int value[];
 	
-	public Elf32_Type(int size,byte[] value,Integer offset){
+	public Elf32_Type(int size,int[] value,Integer offset){
 		this.size=size;
-		this.value=new byte[size];
+		this.value=new int[size];
 		if(offset==null){
 			offset=0;
 		}
@@ -20,11 +20,11 @@ public abstract class Elf32_Type {
 		return size;
 	}
 
-	public byte[] getValue() {
+	public int[] getValue() {
 		return value;
 	}
 
-	protected void setValue(byte[] value) {
+	protected void setValue(int[] value) {
 		for(int i=0;i<this.value.length;i++){
 			this.value[i]=value[i];
 		}
